修改注释信息 解除对py_text的依赖
