+++ src/ndb/operate/node_delete.py
@@ -1,7 +1,7 @@
 # coding=utf-8
 
 '''
-Created on 2014-5-28
+ndb节点删除
 
 @author: Huiyugeng
 '''
