+++ src/ndb/operate/node_select.py
@@ -1,7 +1,7 @@
 # coding=utf-8
 
 '''
-Created on 2014-5-28
+ndb节点查询
 
 @author: Huiyugeng
 '''
