+++ src/ndb/operate/node_update.py
@@ -1,7 +1,7 @@
 # coding=utf-8
 
 '''
-Created on 2014-5-28
+ndb节点更新
 
 @author: Huiyugeng
 '''
